add Python module and the implementation of custom Heap class
Python offers the library heapq for Heap. However, it has limited functionalities. For instance, it only offers min-heap, and any objects (that aren't scalars) are a little complicated to work with. Therefore, as a solution, I built a wrapper class that allows both min- and max-heap, and objects being the elements (e.g. list), as well as the ability to pass in custom comparators